CreateFolder1Test -> POM -> FolderTest, NewItemTest
NewItemTest -> POM
