+++ src/test/java/school/redrover/NewItemTest.java
@@ -1,41 +1,36 @@
 package school.redrover;
 
 import org.openqa.selenium.By;
-import org.openqa.selenium.WebElement;
 import org.testng.Assert;
-import org.testng.annotations.Ignore;
 import org.testng.annotations.Test;
+import school.redrover.model.CreateNewItemPage;
+import school.redrover.model.HomePage;
 import school.redrover.model.ItemPage;
 import school.redrover.runner.BaseTest;
 
+import java.util.List;
+
 public class NewItemTest extends BaseTest {
-
-    public WebElement okButton() {
-        return getDriver().findElement(By.id("ok-button"));
-    }
-
-    public WebElement submitButton(){
-        return getDriver().findElement(By.xpath("//button[@name = 'Submit']"));
-    }
 
     @Test
     public void testAddItem() {
-        new ItemPage(getDriver())
-                .newItemClick()
-                .freestyleProjectClick()
-                .newItemName("NewItemName")
-                .clickButtonOK()
-                .clickLogo();
+        List<String> itemList = new HomePage(getDriver())
+                .clickNewItem()
+                .setItemName("NewItemName")
+                .selectFreestyleAndClickOk()
+                .clickLogo()
+                .getItemList();
 
-        Assert.assertEquals(getDriver().findElement(By.xpath("//span[contains(.,'NewItemName')]")).getText(), "NewItemName");
+        Assert.assertListContainsObject(itemList, "NewItemName", "Item not displayed");
     }
 
     @Test
     public void testGoToNewJobPage() {
-        new ItemPage(getDriver())
-                .newItemClick();
+        String pageHeading = new HomePage(getDriver())
+                .clickNewItem()
+                .getTitleOfNameField();
 
-        Assert.assertTrue(getDriver().findElement(By.xpath("//*[text()='Enter an item name']")).isDisplayed());
+        Assert.assertEquals(pageHeading,"Enter an item name");
     }
 
     @Test
@@ -63,69 +58,32 @@ public class NewItemTest extends BaseTest {
                 .getText(), "NewPipeline");
     }
 
-    @Ignore
     @Test
-    public void testCreateMultiConfigurationProject() {
-        getDriver().findElement(By.xpath("//*[@id='tasks']/div[1]/span/a")).click();
-        getDriver().findElement(By.xpath("//input[@class='jenkins-input']"))
-                .sendKeys("MultiConfigurationProject");
-        getDriver().findElement(By.xpath("//span[contains(text(),  'Multi-configuration project')]")).click();
-        okButton().click();
-        submitButton().click();
-        String result = getDriver().findElement(By.xpath("//h1[@class='matrix-project-headline page-headline']")).getText();
+    public void testOkButtonUsingValidName() {
+        boolean okButtonIsEnabled = new HomePage(getDriver())
+                .clickNewItem()
+                .setItemName("Test Project")
+                .okButtonIsEnabled();
 
-        Assert.assertEquals(result, "Project MultiConfigurationProject");
+        Assert.assertFalse(okButtonIsEnabled);
     }
 
-    @Ignore
     @Test
-    public void testCreateFolder() {
-        getDriver().findElement(By.xpath("//*[@id='tasks']/div[1]/span/a")).click();
-        getDriver().findElement(By.xpath("//input[@class='jenkins-input']"))
-                .sendKeys("Folder");
-        getDriver().findElement(By.xpath("//span[contains(text(),  'Folder')]")).click();
-        okButton().click();
-        submitButton().click();
-        String result = getDriver().findElement(By.xpath("//h1[contains (text(), 'Folder')]")).getText();
+    public void testMessageWhenCreateItemUsingSpecialCharactersInName() {
+        String[] specialCharacters = {"!", "%", "&", "#", "@", "*", "$", "?", "^", "|", "/", "]", "["};
 
-        Assert.assertEquals(result, "Folder");
-    }
+        new HomePage(getDriver())
+                .clickNewItem();
 
-    @Ignore
-    @Test
-    public void testCreateMultibranchPipeline() {
-        getDriver().findElement(By.xpath("//*[@id='tasks']/div[1]/span/a")).click();
-        getDriver().findElement(By.xpath("//input[@class='jenkins-input']"))
-                .sendKeys("MultibranchPipeline");
-        getDriver().findElement(By.xpath("//span[contains(text(),  'Multibranch Pipeline')]")).click();
-        okButton().click();
-        submitButton().click();
-        String result = getDriver().findElement(By.xpath("//h1")).getText();
+        for (String specChar : specialCharacters) {
+            String actualErrorMessage = new CreateNewItemPage(getDriver())
+                    .clearItemNameField()
+                    .setItemName("Fold" + specChar + "erdate")
+                    .getErrorMessage();
 
-        Assert.assertEquals(result, "MultibranchPipeline");
-    }
+            String expectMessage = "» ‘" + specChar + "’ is an unsafe character";
 
-    @Ignore
-    @Test
-    public void testOrganizationFolder() {
-        getDriver().findElement(By.xpath("//*[@id='tasks']/div[1]/span/a")).click();
-        getDriver().findElement(By.xpath("//input[@class='jenkins-input']"))
-                .sendKeys("OrganizationFolder");
-        getDriver().findElement(By.xpath("//span[contains(text(),  'Organization Folder')]")).click();
-        okButton().click();
-        submitButton().click();
-        String result = getDriver().findElement(By.xpath("//h1")).getText();
-
-        Assert.assertEquals(result, "OrganizationFolder");
-    }
-    @Test
-    public void testCheckPage() {
-        getDriver().findElement(By.xpath("//a[@href='newJob']")).click();
-        getDriver().findElement(By.xpath("//input[@class='jenkins-input']"))
-                .sendKeys("Test Project");
-        boolean okButtonIsEnable = getDriver().findElement(By.xpath("//button[@id='ok-button']")).isEnabled();
-        Assert.assertFalse(okButtonIsEnable);
-
-
+            Assert.assertEquals(actualErrorMessage, expectMessage, "Message is not displayed");
+        }
     }
 }
